FIX :: move useTimer to signup.hooks
