+++ packages/maru-hooks/index.ts
@@ -1,5 +1,4 @@
 export { default as useInterval } from './src/useInterval';
 export { default as useDebounceInput } from './src/useDebounceInput';
-export { default as useTimer } from './src/useTimer';
 export { default as useOutsideClick } from './src/useOutsideClick';
 export { default as useBoolean } from './src/useBoolean';
